uwsgi-reload/requirements.txt
Flask==2.3.3
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp-proto-grpc
